Added description of tables.
Added more detail to the comments describing the tables and copied to tabledefinitions for convenience. Cleaned up directory.
